refactor: restructure tests into layered strategy (unit/pairwise/smoke)
Split monolithic generator.test.ts and verify.test.ts into 3 layers:
- Layer A: tests/presets/*.test.ts (8 preset unit tests)
- Layer B: tests/pairwise.test.ts (cross-preset interaction tests)
- Layer C: tests/smoke.test.ts (representative pattern validation)

Extract shared helpers to tests/helpers.ts. Keep resolvePresets tests
and file list snapshots in generator.test.ts. Update verify script to
run smoke tests only.

Closes #74
