+++ package.json
@@ -25,7 +25,7 @@
     "typecheck": "tsc --noEmit",
     "test": "vitest run",
     "test:watch": "vitest",
-    "verify": "vitest run tests/verify.test.ts"
+    "verify": "vitest run tests/smoke.test.ts"
   },
   "engines": {
     "node": ">=22"
