+++ tests/pairwise.test.ts
@@ -1,0 +1,243 @@
+/**
+ * Layer B: Pairwise interaction tests.
+ *
+ * Verifies that important preset combinations produce correct merged output.
+ * Focuses on cross-layer interactions where merge conflicts or deduplication matter.
+ */
+import { describe, expect, it } from "vitest";
+import { generate } from "../src/generator.js";
+import { makeAnswers } from "./helpers.js";
+
+// --- TypeScript + Python ---
+
+describe("pairwise: typescript + python", () => {
+  const answers = makeAnswers({ languages: ["typescript", "python"] });
+  const result = generate(answers);
+
+  it("includes both TypeScript and Python files", () => {
+    expect(result.hasFile("biome.json")).toBe(true);
+    expect(result.hasFile("tsconfig.json")).toBe(true);
+    expect(result.hasFile("pyproject.toml")).toBe(true);
+    expect(result.hasFile("uv.lock")).toBe(true);
+  });
+
+  it("merges both language tools into .mise.toml", () => {
+    const toml = result.readToml(".mise.toml") as Record<string, Record<string, string>>;
+    expect(toml.tools["npm:@biomejs/biome"]).toBe("2");
+    expect(toml.tools.python).toBe("3.12");
+    expect(toml.tools.uv).toBe("0.7");
+  });
+
+  it("merges both language hooks into lefthook.yaml", () => {
+    const hook = result.readYaml("lefthook.yaml") as Record<string, unknown>;
+    const preCommit = hook["pre-commit"] as Record<string, unknown>;
+    const cmds = preCommit.commands as Record<string, unknown>;
+    expect(cmds.biome).toBeDefined();
+    expect(cmds["ruff-format"]).toBeDefined();
+  });
+
+  it("has CI steps from both presets", () => {
+    const ci = result.readYaml(".github/workflows/ci.yaml") as Record<string, unknown>;
+    const jobs = ci.jobs as Record<string, Record<string, unknown>>;
+    const steps = jobs["lint-and-check"].steps as Array<Record<string, unknown>>;
+    const stepNames = steps.map((s) => s.name);
+    expect(stepNames).toContain("Lint (Biome)");
+    expect(stepNames).toContain("Lint (Ruff)");
+    expect(stepNames).toContain("Test");
+    expect(stepNames).toContain("Test (pytest)");
+  });
+
+  it("expands CLAUDE.md with both language sections", () => {
+    const claude = result.readText("CLAUDE.md");
+    expect(claude).toContain("TypeScript");
+    expect(claude).toContain("Python");
+    expect(claude).not.toContain("<!-- SECTION:");
+  });
+
+  it("expands git-workflow.md with both Biome/Ruff and typecheck/mypy", () => {
+    const gw = result.readText(".claude/rules/git-workflow.md");
+    expect(gw).toContain("Biome");
+    expect(gw).toContain("Ruff");
+    expect(gw).toContain("typecheck");
+    expect(gw).toContain("mypy");
+    expect(gw).not.toContain("<!-- SECTION:");
+  });
+
+  it("lint:all includes both TypeScript and Python lint scripts", () => {
+    const pkg = result.readJson("package.json") as Record<string, unknown>;
+    const scripts = pkg.scripts as Record<string, string>;
+    expect(scripts["lint:all"]).toContain("pnpm run lint");
+    expect(scripts["lint:all"]).toContain("pnpm run typecheck");
+    expect(scripts["lint:all"]).toContain("pnpm run lint:python");
+    expect(scripts["lint:all"]).toContain("pnpm run lint:mypy");
+    expect(scripts["lint:all"]).toContain("pnpm run lint:secrets");
+  });
+
+  it("merges both language settings into VSCode settings", () => {
+    const settings = result.readJson(".vscode/settings.json") as Record<string, unknown>;
+    expect(settings["editor.defaultFormatter"]).toBe("biomejs.biome");
+    expect(settings["[typescript]"]).toBeDefined();
+    expect(settings["mypy-type-checker.importStrategy"]).toBe("fromEnvironment");
+    expect(settings["[python]"]).toBeDefined();
+  });
+
+  it("merges both language extensions into VSCode extensions", () => {
+    const ext = result.readJson(".vscode/extensions.json") as Record<string, string[]>;
+    expect(ext.recommendations).toContain("biomejs.biome");
+    expect(ext.recommendations).toContain("charliermarsh.ruff");
+    expect(ext.recommendations).toContain("ms-python.python");
+    expect(ext.recommendations).toContain("EditorConfig.EditorConfig");
+  });
+
+  it("merges both language extensions and mounts into devcontainer", () => {
+    const dc = result.readJson(".devcontainer/devcontainer.json") as Record<string, unknown>;
+    const customizations = dc.customizations as Record<string, Record<string, string[]>>;
+    expect(customizations.vscode.extensions).toContain("biomejs.biome");
+    expect(customizations.vscode.extensions).toContain("charliermarsh.ruff");
+    expect(customizations.vscode.extensions).toContain("ms-python.python");
+    const mounts = dc.mounts as string[];
+    expect(mounts.some((m: string) => m.includes("uv-cache"))).toBe(true);
+    expect(mounts.some((m: string) => m.includes("pnpm-store"))).toBe(true);
+  });
+});
+
+// --- Python + Terraform ---
+
+describe("pairwise: python + terraform", () => {
+  const answers = makeAnswers({ languages: ["python"], clouds: ["aws"], iac: ["terraform"] });
+  const result = generate(answers);
+
+  it("includes Python and Terraform files", () => {
+    expect(result.hasFile("pyproject.toml")).toBe(true);
+    expect(result.hasFile(".tflint.hcl")).toBe(true);
+  });
+
+  it("does not include TypeScript files", () => {
+    expect(result.hasFile("biome.json")).toBe(false);
+    expect(result.hasFile("tsconfig.json")).toBe(false);
+  });
+
+  it("merges both preset tools into .mise.toml", () => {
+    const toml = result.readToml(".mise.toml") as Record<string, Record<string, string>>;
+    expect(toml.tools.python).toBe("3.12");
+    expect(toml.tools.terraform).toBe("1");
+    expect(toml.tools.tflint).toBe("0.55");
+  });
+
+  it("has CI steps from both presets", () => {
+    const ci = result.readYaml(".github/workflows/ci.yaml") as Record<string, unknown>;
+    const jobs = ci.jobs as Record<string, Record<string, unknown>>;
+    const steps = jobs["lint-and-check"].steps as Array<Record<string, unknown>>;
+    const stepNames = steps.map((s) => s.name);
+    expect(stepNames).toContain("Lint (Ruff)");
+    expect(stepNames).toContain("Lint (Terraform)");
+  });
+});
+
+// --- Python + CloudFormation ---
+
+describe("pairwise: python + cloudformation", () => {
+  const answers = makeAnswers({ languages: ["python"], clouds: ["aws"], iac: ["cloudformation"] });
+  const result = generate(answers);
+
+  it("includes Python and CloudFormation files", () => {
+    expect(result.hasFile("pyproject.toml")).toBe(true);
+    expect(result.hasFile("infra/template.yaml")).toBe(true);
+    expect(result.hasFile(".cfnlintrc.yaml")).toBe(true);
+  });
+
+  it("does not include TypeScript or Terraform files", () => {
+    expect(result.hasFile("biome.json")).toBe(false);
+    expect(result.hasFile(".tflint.hcl")).toBe(false);
+  });
+
+  it("merges both preset tools into .mise.toml", () => {
+    const toml = result.readToml(".mise.toml") as Record<string, Record<string, string>>;
+    expect(toml.tools.python).toBe("3.12");
+    expect(toml.tools["pipx:cfn-lint"]).toBe("1");
+  });
+
+  it("has CI steps from both presets", () => {
+    const ci = result.readYaml(".github/workflows/ci.yaml") as Record<string, unknown>;
+    const jobs = ci.jobs as Record<string, Record<string, unknown>>;
+    const steps = jobs["lint-and-check"].steps as Array<Record<string, unknown>>;
+    const stepNames = steps.map((s) => s.name);
+    expect(stepNames).toContain("Lint (Ruff)");
+    expect(stepNames).toContain("Lint (cfn-lint)");
+  });
+
+  it("replaces {{projectName}} in CloudFormation template", () => {
+    const template = result.readText("infra/template.yaml");
+    expect(template).toContain("test-app");
+    expect(template).not.toContain("{{projectName}}");
+  });
+});
+
+// --- CDK + CloudFormation (cfn-lint deduplication) ---
+
+describe("pairwise: cdk + cloudformation (cfn-lint dedup)", () => {
+  it("cfn-lint appears only once when both CDK and CloudFormation are selected", () => {
+    const result = generate(
+      makeAnswers({
+        languages: ["typescript"],
+        clouds: ["aws"],
+        iac: ["cdk", "cloudformation"],
+      }),
+    );
+    const ci = result.readText(".github/workflows/ci.yaml");
+    const matches = ci.match(/Lint \(cfn-lint\)/g) ?? [];
+    expect(matches.length, "cfn-lint should appear exactly once in CI").toBe(1);
+  });
+});
+
+// --- React + CDK (web/ + infra/ coexistence) ---
+
+describe("pairwise: react + cdk", () => {
+  const answers = makeAnswers({ frontend: "react", clouds: ["aws"], iac: ["cdk"] });
+  const result = generate(answers);
+
+  it("includes both web/ and infra/ directories", () => {
+    expect(result.hasFile("web/vite.config.ts")).toBe(true);
+    expect(result.hasFile("web/package.json")).toBe(true);
+    expect(result.hasFile("infra/bin/app.ts")).toBe(true);
+    expect(result.hasFile("infra/package.json")).toBe(true);
+  });
+
+  it("workspace includes web only (infra uses independent npm)", () => {
+    const workspace = result.readText("pnpm-workspace.yaml");
+    expect(workspace).toContain("- web");
+    expect(workspace).not.toContain("- infra");
+  });
+
+  it("root package.json has both orchestration and CDK scripts", () => {
+    const pkg = result.readJson("package.json") as Record<string, unknown>;
+    const scripts = pkg.scripts as Record<string, string>;
+    expect(scripts["build:web"]).toBe("pnpm --filter web build");
+    expect(scripts["cdk:synth"]).toContain("cdk synth");
+  });
+});
+
+// --- React + Python (web/ + Python tools coexistence) ---
+
+describe("pairwise: react + python", () => {
+  const answers = makeAnswers({ frontend: "react", languages: ["python"] });
+  const result = generate(answers);
+
+  it("includes web/ and Python files", () => {
+    expect(result.hasFile("web/vite.config.ts")).toBe(true);
+    expect(result.hasFile("pyproject.toml")).toBe(true);
+  });
+
+  it("merges both TypeScript and Python tools (React forces TypeScript)", () => {
+    const toml = result.readToml(".mise.toml") as Record<string, Record<string, string>>;
+    expect(toml.tools["npm:@biomejs/biome"]).toBe("2");
+    expect(toml.tools.python).toBe("3.12");
+  });
+
+  it("lint:all includes both TypeScript and Python lint scripts", () => {
+    const pkg = result.readJson("package.json") as Record<string, unknown>;
+    const scripts = pkg.scripts as Record<string, string>;
+    expect(scripts["lint:all"]).toContain("pnpm run lint");
+    expect(scripts["lint:all"]).toContain("pnpm run lint:python");
+  });
+});
